# Add vectorized ParallelFor function and helper macros