--- a/cpp/open3d/Macro.h
+++ b/cpp/open3d/Macro.h
@@ -28,9 +28,6 @@
 
 #include <cassert>
 
-#define OPEN3D_CONCAT_IMPL(s1, s2) s1##s2
-#define OPEN3D_CONCAT(s1, s2) OPEN3D_CONCAT_IMPL(s1, s2)
-
 // https://gcc.gnu.org/wiki/Visibility updated to use C++11 attribute syntax
 #if defined(_WIN32) || defined(__CYGWIN__)
 #define OPEN3D_DLL_IMPORT __declspec(dllimport)