--- a/cpp/open3d/utility/Helper.isph
+++ b/cpp/open3d/utility/Helper.isph
@@ -23,30 +23,35 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
-// Copyright 2018 Google Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
 
-#include <benchmark/benchmark.h>
+#pragma once
 
-#include "open3d/Open3D.h"
+// This C/C++ header is compatible with ISPC.
+// Variadic macros, i.e. ellipsis (...) and __VA_ARGS__ are also supported.
+#include "open3d/utility/Preprocessor.h"
 
-int main(int argc, char** argv) {
-    open3d::utility::CPUInfo::GetInstance().Print();
-    open3d::utility::ISAInfo::GetInstance().Print();
-    benchmark::Initialize(&argc, argv);
-    if (benchmark::ReportUnrecognizedArguments(argc, argv)) {
-        return 1;
-    }
-    benchmark::RunSpecifiedBenchmarks();
-}
+/// CHAR_BIT
+///
+/// The number of bits per byte.
+#define CHAR_BIT 8
+
+/// Enable C/C++ fixed size integer types.
+typedef uint8 uint8_t;
+typedef int8 int8_t;
+typedef uint16 uint16_t;
+typedef int16 int16_t;
+typedef uint32 uint32_t;
+typedef int32 int32_t;
+typedef uint64 uint64_t;
+typedef int64 int64_t;
+
+/// OPEN3D_ENSURE_EXPORTED(StructOrEnum)
+///
+/// Ensures that the provided struct or enum is exported to the C++ side.
+///
+/// This is useful if no other exported function directly or indirectly
+/// references the struct or enum via its arguments.
+#define OPEN3D_ENSURE_EXPORTED(StructOrEnum) \
+    export void OPEN3D_CONCAT(               \
+            DummyToEnsureExported_,          \
+            OPEN3D_CONCAT(StructOrEnum, _))(uniform StructOrEnum * uniform) {}