#### Highlights:

- `ParallelFor` with additional argument to call the vectorized code path. Falls back to other `ParallelFor` if `BUILD_ISPC_MODULE` is `OFF`.
- `OPEN3D_VECTORIZED` and `OPEN3D_TEMPLATE_VECTORIZED` macros to automatically create a lambda function calling the specified ISPC kernel. Only work for up to 9 additional parameters.
- `OPEN3D_EXPORT_VECTORIZED` and `OPEN3D_EXPORT_TEMPLATE_VECTORIZED` macros to define the ISPC kernels to be called by `ParallelFor`. Only work for up to 9 additional parameters.
- `OPEN3D_INSTANTIATE_TEMPLATE` and `OPEN3D_INSTANTIATE_TEMPLATE_WITH_BOOL` macros to emulate template functions and their instantiation in ISPC code.
- Unit tests to cover the introduced functions and macros.

#### Benchmarks (Intel i7-10750H):

Command: `./bin/benchmarks --benchmark_filter=ParallelFor`

```
[Open3D INFO] CPUInfo: 6 cores, 12 threads.
[Open3D INFO] ISAInfo: AVX2 instruction set used in ISPC code.
2021-09-14T11:44:43+02:00
Running ./bin/benchmarks
Run on (12 X 4727.5 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 1.76, 1.27, 1.13
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
ParallelForScalar/CPU100                 1.27 us         1.27 us       542632
ParallelForScalar/CPU1000                1.26 us         1.26 us       533485
ParallelForScalar/CPU10000               2.14 us         2.14 us       329166
ParallelForScalar/CPU100000              11.4 us         11.4 us        55214
ParallelForScalar/CPU1000000              105 us          105 us         6769
ParallelForScalar/CPU10000000            3714 us         3714 us          188
ParallelForScalar/CPU100000000          39998 us        39995 us           18
ParallelForVectorized/CPU100             1.38 us         1.38 us       510147
ParallelForVectorized/CPU1000            1.40 us         1.40 us       496748
ParallelForVectorized/CPU10000           1.51 us         1.51 us       462849
ParallelForVectorized/CPU100000          3.59 us         3.59 us       197242
ParallelForVectorized/CPU1000000         42.2 us         42.2 us        16401
ParallelForVectorized/CPU10000000        3659 us         3658 us          194
ParallelForVectorized/CPU100000000      39473 us        39473 us           18
```

Command: `OPENMP_NUM_THREADS=1 ./bin/benchmarks --benchmark_filter=ParallelFor`

```
[Open3D INFO] CPUInfo: 6 cores, 12 threads.
[Open3D INFO] ISAInfo: AVX2 instruction set used in ISPC code.
2021-09-14T12:08:34+02:00
Running ./bin/benchmarks
Run on (12 X 4795.83 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
Load Average: 0.61, 0.46, 0.57
-----------------------------------------------------------------------------
Benchmark                                   Time             CPU   Iterations
-----------------------------------------------------------------------------
ParallelForScalar/CPU100                0.554 us        0.554 us      1252981
ParallelForScalar/CPU1000               0.940 us        0.940 us       733749
ParallelForScalar/CPU10000               4.97 us         4.97 us       139518
ParallelForScalar/CPU100000              44.6 us         44.6 us        15436
ParallelForScalar/CPU1000000              444 us          444 us         1570
ParallelForScalar/CPU10000000            5207 us         5207 us          130
ParallelForScalar/CPU100000000          52967 us        52968 us           13
ParallelForVectorized/CPU100            0.597 us        0.597 us      1170738
ParallelForVectorized/CPU1000           0.662 us        0.662 us      1039586
ParallelForVectorized/CPU10000           1.40 us         1.40 us       502559
ParallelForVectorized/CPU100000          14.0 us         14.0 us        50152
ParallelForVectorized/CPU1000000          145 us          145 us         4678
ParallelForVectorized/CPU10000000        3924 us         3924 us          178
ParallelForVectorized/CPU100000000      42717 us        42718 us           16
```
#### Observations:
- Performance improvement becomes more significant with less cores used.
- Array size in relation to CPU Cache seems to have significant positive effect:
   - **Small**: Compute-bound, overhead exceeds actual work.
   - **Mid-sized**: Compute-bound, 2-4x speed improvements with `AVX2`.
   - **Large**: Memory-bound, acceleration less significant.
- More compute-intense kernels could benefit more from vectorization.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/isl-org/open3d/4039)
<!-- Reviewable:end -->