@stotko 2021-09-15
I have added two new helper functions on the ISPC side:

- `OPEN3D_ENSURE_EXPORTED()`: Used to make a `struct` or `enum` visible in in the corresponding `xxx_ispc.h` header independent of whether it is used in another `export`ed function.
- Typedefs for the fixed size integer types, e.g. `int64` -> `int64_t`: This makes the syntax look much more like C/C++.

Since common, standard-like functionality is currently handled in `Helper.h`, putting above functionality into `Helper.isph` seemed like a natural choice.

@yxlao @ssheorey Please have a look again at this PR. Functionality-wise it should be complete now and CI should pass as well.