@yxlao 2021-09-16
@stotko I am observing a strange behavior in my system.
1. Clean the `build` directory, and build from scratch. For example, configure and run `make test`.
2. Modify any CMake file, e.g. change comment formatting in the root `CMakeLists.txt`. 
3. Rerun `make test`.
   - The compilation eventually fails at:
     ```
     [ 43%] Building ISPC object cpp/open3d/utility/CMakeFiles/utility.dir/ISAInfo.ispc.o
     Consolidate compiler generated dependencies of target core
     Error: Unknown option "-pthread"
     ```
   - In my case, `ccache` is also no longer working properly, as it treats everything as recompilation. Perhaps this is related to the issue above (wrong compilation flag).

This is true both in `master` and in this PR.