@stotko 2021-09-17
> @stotko I am observing a strange behavior:
> 
>     1. Clean the `build` directory, and build from scratch. For example, configure and run `make test`.
> 
>     2. Modify any CMake file, e.g. change comment formatting in the root `CMakeLists.txt`.
> 
>     3. Rerun `make test`.
>        
>        * The compilation eventually fails at:
>          ```
>          [ 43%] Building ISPC object cpp/open3d/utility/CMakeFiles/utility.dir/ISAInfo.ispc.o
>          Consolidate compiler generated dependencies of target core
>          Error: Unknown option "-pthread"
>          ```
>        * In my case, `ccache` is also no longer working properly, as it treats everything as recompilation. Perhaps this is related to the compilation flag issue above.
> 
> 
> This is true both in `master` and in this PR. It means that every time we update cmake, we have to clean the `build` directory and rebuild.

Interesting. It seems like the "state" of the project, i.e. the values of some CMake variables, always changes in such cases. The build error you mentioned is likely just a part of that state change where the `Threads::Threads` target changed its properties for some reason and ignores our setting of `THREADS_PREFER_PTHREAD_FLAG`.

Since this also happens on `master` (for a longer time I think), the root cause is unrelated to this PR and needs further investigation. In the meantime, we could consider forcing the legacy ISPC support which does not suffer from the `-pthread` flag issue due to its several limitations. Other options could be to report this issue upstream to ISPC and/or CMake to "fix" first-class ISPC language support.