@stotko 2021-09-17
> > Interesting. It seems like the "state" of the project, i.e. the values of some CMake variables, always changes in such cases. The build error you mentioned is likely just a part of that state change where the `Threads::Threads` target changed its properties for some reason and ignores our setting of `THREADS_PREFER_PTHREAD_FLAG`.
> > Since this also happens on `master` (for a longer time I think), the root cause is unrelated to this PR and needs further investigation. In the meantime, we could consider forcing the legacy ISPC support which does not suffer from the `-pthread` flag issue due to its several limitations. Other options could be to report this issue upstream to ISPC and/or CMake to "fix" first-class ISPC language support.
> 
> The issue happens to me after the first ISPC PR #3996. Did you observe similar issues?

I observed the general issue of triggering recompilation of every source file after a change in a `CMakeLists.txt` file. The `-pthread` issue is likely just another thing that "gets broken" for some reason. The ISPC support itself should be fine. Instead, I suspect that it could related to `ccache` itself and/or `CMAKE_<LANG>_COMPILER_LAUNCHER`, but I do not have any concrete proofs for that.